+++ src/components/common/index.ts
@@ -2,4 +2,4 @@ export * from './Spinner'
 export * from './Text'
 export * from './Input'
 export * from './Button'
-export * from './Placeholder'
+export * from './Placehold'
